update bugs program donor
