+++ project/resources/views/tr/program/js/detail-edit/donor_old.blade.php
@@ -5,7 +5,6 @@
         var placeholder = "{{ __('cruds.program.donor.select') }}";
         var url_program_pendonor = "{{ route('api.pendonor.data', ':id') }}".replace(':id', programId);
         $(document).ready(function() {
-
             $.ajax({
                 url: url_program_pendonor,
                 method: 'get',
@@ -37,7 +36,7 @@
                             ajax: {
                                 url: data_donor,
                                 method: 'GET',
-                                delay: 500,
+                                delay: 10000,
                                 processResults: function(data) {
                                     return {
                                         results: data.map(function(item) {
@@ -61,32 +60,14 @@
                         $('#donor').val(selected_pendonor).trigger('change');
                     }, 1000);
                 },
-                error: function(xhr, status, error) {
-                    console.error("Error fetching program pendonors:", status, error);
-                    // Optionally show user-friendly error message
-                    Swal.fire({
-                        icon: 'error',
-                        title: 'Oops...',
-                        text: 'Failed to load donor information'
-                    });
+                error: function(xhr) {
+                    console.log("AJAX DATA ERROR: ".xhr)
                 }
             });
-
             $('#donor').change(function() {
                 var selected = $(this).val();
-                var existingDonors = new Set();
-
-                $('#pendonor-container .row').each(function() {
-                    var existingDonorId = $(this).find('input[name="pendonor_id[]"]').val();
-                    existingDonors.add(existingDonorId);
-                });
-
                 $('#pendonor-container').empty();
-
-
                 selected.forEach(function(pendonor_id) {
-                    if (existingDonors.has(pendonor_id)) return;
-
                     if ($('#pendonor-container').find(`#pendonor-${pendonor_id}`).length === 0) {
                         var data_pendonor = '{{ route('api.search.pendonor', ':id') }}'.replace(
                                 ':id', pendonor_id) + '?program_id=' +
@@ -96,7 +77,7 @@
                             type: 'GET',
                             url: data_pendonor,
                             dataType: 'json',
-                            delay: 500,
+                            delay: 1000,
                             success: function(data) {
                                 setTimeout(() => {
                                     if (data && Array.isArray(data)) {
@@ -129,7 +110,7 @@
                                                         <div class="col-lg-3 form-group">
                                                             <div class="input-group">
                                                                 <label for="nilaidonasi-${pendonor.id}" class="input-group small mb-0">{{ __('cruds.program.donor.val') }}</label>
-                                                                <input type="text" id="nilaidonasi-${pendonor.id}" name="nilaidonasi[]" class="form-control currency nilaidonasi" value="${pendonor.nilaidonasi || 0}">
+                                                                <input type="text" id="nilaidonasi-${pendonor.id}" name="nilaidonasi[]" class="form-control currency" value="${pendonor.nilaidonasi || 0}">
                                                                 <span class="input-group-append">
                                                                     <button type="button" class="btn btn-danger form-control remove-pendonor nilaidonasi btn-flat" data-target="${containerId}"><i class="bi bi-trash"></i></button>
                                                                 </span>
@@ -137,13 +118,27 @@
                                                         </div>
                                                     </div>
                                                 `);
-                                            // Initialize AutoNumeric for donation value
-                                            new AutoNumeric(`#nilaidonasi-${pendonor.id}`, {
-                                                digitGroupSeparator: '.',
-                                                decimalCharacter: ',',
-                                                currencySymbol: 'Rp ',
-                                                modifyValueOnWheel: false
-                                            });
+                                            var nilaidonasiElement =
+                                                `#nilaidonasi-${pendonor.id}`;
+                                            if (!AutoNumeric
+                                                .getAutoNumericElement(
+                                                    nilaidonasiElement)
+                                            ) {
+                                                // console.log(`Initializing AutoNumeric for ${nilaidonasiElement} with value: ${pendonor.nilaidonasi}`);
+                                                new AutoNumeric(
+                                                    nilaidonasiElement, {
+                                                        digitGroupSeparator: '.',
+                                                        decimalCharacter: ',',
+                                                        currencySymbol: 'Rp ',
+                                                        modifyValueOnWheel: false
+                                                    });
+
+                                                AutoNumeric
+                                                    .getAutoNumericElement(
+                                                        nilaidonasiElement
+                                                    ).set(pendonor
+                                                        .nilaidonasi);
+                                            }
                                         });
                                     } else {
                                         console.error('Invalid data format',
@@ -160,133 +155,11 @@
                 });
             });
 
-            let rowCounter = 1;
-            function initializeSelect2(rowElement) {
-                rowElement.find('select[name="pendonor[]"]').select2({
-
-                    placeholder:  "Pilih Pendonor",
-                    width: '100%',
-                    allowClear: true,
-                    ajax: {
-                        url: "{{ route('api.program.donor') }}",
-                        dataType: 'json',
-                        delay: 250,
-                        data: function(params) {
-                            return {
-                                search: params.term,
-                                page: params.page || 1
-                            };
-                        },
-                        processResults: function(data, params) {
-                            params.page = params.page || 1;
-
-                            return {
-                                results: data.data.map(function(item) {
-                                    return {
-                                        id: item.id,
-                                        text: item.nama
-                                    };
-                                }),
-                                pagination: {
-                                    more: data.current_page < data.last_page
-                                }
-                            };
-                        },
-                        cache: true
-                    },
-                });
-
-                // Add change event to populate additional fields
-                rowElement.find('select[name="pendonor[]"]').on('change', function() {
-                    const selectedId = $(this).val();
-                    const rowContainer = $(this).closest('.row');
-                    let url_search = "{{ route('api.search.pendonor', ':id') }}".replace(':id', selectedId);
-                    // Fetch pendonor details
-                    $.ajax({
-
-                        url: url_search,
-                        method: 'GET',
-                        success: function(data) {
-                            if (data && data.length > 0) {
-                                const pendonor = data[0];
-
-                                // Populate hidden inputs
-                                rowContainer.find('input[name="pendonor_id[]"]').val(pendonor.id);
-                                rowContainer.find('input[name="program_id[]"]').val(pendonor.id);
-
-                                // Populate other fields
-                                rowContainer.find('input[name="email"]').val(pendonor.email || '');
-                                rowContainer.find('input[name="phone"]').val(pendonor.phone || '');
-
-                                // Initialize AutoNumeric for donation value
-                                new AutoNumeric(rowContainer.find('input[name="nilaidonasi[]"]')[0], {
-                                    digitGroupSeparator: '.',
-                                    decimalCharacter: ',',
-                                    currencySymbol: 'Rp ',
-                                    modifyValueOnWheel: false
-                                });
-                            }
-                        },
-                        error: function(xhr) {
-                            console.error('Error fetching pendonor details', xhr);
-                        }
-                    });
-                });
-            }
-             // Add new row button click handler
-            $('#addDonasiButton').on('click', function() {
-                rowCounter++; // Increment counter
-                // Clone the template row
-                const newRow = $(`
-                    <div class="row container-pendonor" id="pendonor-row-${rowCounter}">
-                        <div class="col-lg-3 form-group">
-                            <div class="input-group">
-                                <label for="nama_pendonor-${rowCounter}" class="input-group small mb-0">{{ __('cruds.program.donor.nama') }}</label>
-                                <input type="hidden" name="pendonor_id[]">
-                                <input type="hidden" name="program_id[]">
-                                <select name="pendonor[]" class="form-control select2"></select>
-                            </div>
-                        </div>
-                        <div class="col-lg-3 form-group">
-                            <div class="input-group">
-                                <label for="email-${rowCounter}" class="input-group small mb-0">{{ __('cruds.program.donor.email') }}</label>
-                                <input type="text" id="email-${rowCounter}" name="email" class="form-control" readonly>
-                            </div>
-                        </div>
-                        <div class="col-lg-2 form-group">
-                            <div class="input-group">
-                                <label for="phone-${rowCounter}" class="input-group small mb-0">{{ __('cruds.program.donor.ph') }}</label>
-                                <input type="text" id="phone-${rowCounter}" name="phone" class="form-control" readonly>
-                            </div>
-                        </div>
-                        <div class="col-lg-3 form-group">
-                            <div class="input-group">
-                                <label for="nilaidonasi-${rowCounter}" class="input-group small mb-0">{{ __('cruds.program.donor.val') }}</label>
-                                <input type="text" id="nilaidonasi-${rowCounter}" name="nilaidonasi[]" class="form-control currency nilaidonasi-value">
-                                <span class="input-group-append">
-                                    <button type="button" class="btn btn-danger form-control remove-pendonor nilaidonasi btn-flat" data-target="pendonor-row-${rowCounter}">
-                                        <i class="bi bi-trash"></i>
-                                    </button>
-                                </span>
-                            </div>
-                        </div>
-                    </div>
-                `);
-
-                $('#pendonor-container').append(newRow);
-
-
-                initializeSelect2(newRow);
-            });
-
-            // Remove row button handler (using event delegation)
             $(document).on('click', '.remove-pendonor', function() {
-                const targetId = $(this).data('target');
+                let targetId = $(this).data('target');
                 $('#' + targetId).remove();
             });
 
-            // Initialize the first row's select2 if it exists
-            initializeSelect2($('.container-pendonor').first());
 
 
         });
